APIs/Oauth/requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncpg==0.22.0
attrs==20.3.0
authclient==1.0
certifi==2020.12.5
chardet==4.0.0
ecdsa==0.14.1
enum-compat==0.0.3
future==0.18.2
greenlet==1.0.0
httptools==0.2.0
idna==2.10
intuit-oauth==1.2.3
multidict==5.1.0
oauthlib==3.1.0
pyasn1==0.4.8
pycparser==2.18
python-dateutil==2.8.1
python-jose==3.2.0
python-quickbooks==0.8.4
rauth==0.7.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
sanic==21.3.4
Sanic-Cors==1.0.0
sanic-plugin-toolkit==1.0.1
sanic-routing==0.6.2
simplejson==3.17.2
six==1.15.0
SQLAlchemy==1.4.11
typing-extensions==3.7.4.3
ujson==4.0.2
urllib3==1.26.4
uvloop==0.15.2
websockets==8.1
yarl==1.6.3
